refactor(landing): clarify src architecture and naming
- proto/ -> scenes/ ; Deck -> SceneStack, useDeckSubState -> useSceneSubState
- pages/auth/ -> auth/ (no other pages exist)
- lib/web3/ folded into lib/services/ (WalletProvider, contractErrors)
- PlateA -> TopicsRadar
- Instruments.tsx (1073 l.) -> IsoStack.tsx (~420 l.); drop unused
  TopicsIntentions/ThreeCircles dead exports
- AnimatedHexSplit extracted from App.tsx -> components/HexSplit.tsx,
  co-located with its module.css
- Drop 1-line useWalletConnection.ts re-export; consumers import the
  hook directly from lib/services/WalletProvider
- OAuth routes in main.tsx (10 duplicates) replaced by flatMap over
  OAUTH_PLATFORMS; OAuthPlatform now derived via typeof[number]
