+++ apps/landing/src/auth/AuthPage.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState, useCallback, useRef } from 'react'
 import { useWallets } from '@privy-io/react-auth'
-import { useWalletConnection } from '../../lib/web3/PrivyContext'
+import { useWalletConnection } from '../lib/services/WalletProvider'
 import { sendToExtension, DEFAULT_EXTENSION_ID } from './oauthConfig'
 import styles from './auth.module.css'
