--- apps/landing/src/components/Instruments.module.css
@@ -1,13 +1,0 @@
-/**
- * Instruments — shared SVG-display class for the PLATE.A / PLATE.C /
- * PLATE.D micrographics. Layout is owned by the host Plate body via
- * `instrument="iso|circles|topics"` in Plate.module.css.
- */
-
-.svg {
-  width: 100%;
-  height: 100%;
-  display: block;
-  max-width: 100%;
-  max-height: 100%;
-}
