+++ apps/landing/src/components/IsoStack.module.css
@@ -1,0 +1,12 @@
+/**
+ * IsoStack — SVG container fills its host slot. The actual layout
+ * (grid placement, max sizing) lives on the parent slide body.
+ */
+
+.svg {
+  width: 100%;
+  height: 100%;
+  display: block;
+  max-width: 100%;
+  max-height: 100%;
+}
