--- apps/landing/src/hooks/useWalletConnection.ts
@@ -1,1 +1,0 @@
-export { useWalletConnection } from '../lib/web3/PrivyContext'
