+++ apps/landing/src/scenes/types.ts
@@ -1,0 +1,7 @@
+export type SlideBg = 'peach' | 'dark'
+
+/** A discrete scene index. Goes from 0 (first scene) to N-1 (last). */
+export type SceneState = number
+
+/** Direction of intent inferred from a wheel/touch/key event. */
+export type Intent = -1 | 0 | 1
